requirements.txt
streamlit
snowflake-connector-python
pandas
openai
matplotlib
scikit-learn
python-dotenv